+++ functions/public/Get-PSProjectStatus.ps1
@@ -1,5 +1,6 @@
 Function Get-PSProjectStatus {
     [cmdletbinding()]
+    [alias("gpstat")]
     [OutputType("PSProject")]
     Param(
         [Parameter(
@@ -16,8 +17,6 @@ Function Get-PSProjectStatus {
         Write-Verbose "Starting $($MyInvocation.MyCommand)"
     }
     Process {
-
-
         $json = Join-Path (Convert-Path $path) -ChildPath psproject.json
         if (Test-Path $json) {
             Write-Verbose "Getting project status from $json"
@@ -27,8 +26,19 @@ Function Get-PSProjectStatus {
             #get property names from the class
             $properties = $psproject.psobject.properties.name | Where-Object { $_ -ne "Age" }
             foreach ($property in $properties) {
-                Write-Verbose "Adding property $property"
-                $psproject.$property = $in.$property
+                if ($property -eq 'RemoteRepository') {
+                    Write-Verbose "Creating remote repository information"
+                    $remote = @()
+                    foreach ($repo in $in.RemoteRepository) {
+                        $remote += [PSProjectRemote]::new($repo.name, $repo.url, $repo.mode)
+                    }
+                    $psproject.RemoteRepository = $remote
+
+                }
+                else {
+                    Write-Verbose "Adding property $property"
+                    $psproject.$property = $in.$property
+                }
             }
             #datetime is stored as a UTC value so convert it to local
             $psproject.lastUpdate = $psproject.lastUpdate.tolocalTime()